cleansing of original sin
Our father, ForgedLives, thou art thy Neuron. Hallowed be his name. Thy minecraft come, thy wins be done. On the Overworld as it is in the End. Give us this day, our daily porkchop, and forgive our trespasses. As we forgive those who has hit the nae-nae. And lead us to not to Umamusume, but deliver us from the Fat one. Amen.

Original has been purged.